Longs everywhere

+++ src/main/scala/com/github/dafutils/chatroom/http/model/AddMessages.scala
@@ -3,6 +3,6 @@ package com.github.dafutils.chatroom.http.model
 import scala.collection.immutable.Seq
 
 case class AddMessages(
-  chatRoomId: Int,
+  chatRoomId: Long,
   messages: Seq[ChatroomMessage]                      
 )
