+++ src/main/scala/com/github/dafutils/chatroom/http/model/ChatroomMessage.scala
@@ -3,7 +3,7 @@ package com.github.dafutils.chatroom.http.model
 import uk.gov.hmrc.emailaddress.EmailAddress
 
 case class ChatroomMessage(
-  index: Int,
+  index: Long,
   timestamp: Long,
   author: EmailAddress,
   message: String                        
@@ -12,7 +12,7 @@ case class ChatroomMessage(
 }
 
 case class ChatroomMessageWithStats(
-  chatroomId: Int,
+  chatroomId: Long,
   timeSincePreviousMessage: Long,
   message: ChatroomMessage,
 )
