+++ src/main/scala/com/github/dafutils/chatroom/http/model/Pauses.scala
@@ -1,3 +1,3 @@
 package com.github.dafutils.chatroom.http.model
 
-case class Pauses(count: Int)
+case class Pauses(count: Long)
